fix: verify request based on public url
When running behind a reverse proxy hosting
the service under a webroot other than `/`
causes 401 error codes due to mismatching
Message Authentication Codes (MACs).

Changes made in this commit allow users
hosting the sync server behind a reverse proxy
to specify the `public_url` of their service
in order to correct this behaviour.

In doing so, changes made in this commit fix #1217, #1649